chore: remove foreign call logic from `noirc_printable_type`
