+++ tooling/nargo/src/foreign_calls/rpc.rs
@@ -2,10 +2,9 @@ use std::path::PathBuf;
 
 use acvm::{acir::brillig::ForeignCallResult, pwg::ForeignCallWaitInfo, AcirField};
 use jsonrpc::{arg as build_json_rpc_arg, minreq_http::Builder, Client};
-use noirc_printable_type::ForeignCallError;
 use serde::{Deserialize, Serialize};
 
-use super::ForeignCallExecutor;
+use super::{ForeignCallExecutor, ForeignCallError};
 
 #[derive(Debug)]
 pub(super) struct RPCForeignCallExecutor {
